+++ packages/edit/src/components/Edit.vue
@@ -1,135 +1,151 @@
 <template>
-  <VForm
-    ref="form"
-    class="tce-container"
-    validate-on="submit"
-    @submit.prevent="save"
+  <QuestionContainer
+    v-bind="{
+      type: manifest.name,
+      icon: manifest.ui.icon,
+      embedTypes,
+      elementData,
+      isDirty,
+      isDisabled,
+    }"
+    show-feedback
+    @cancel="updateData(element.data)"
+    @save="save"
+    @update="updateData($event)"
   >
-    <VTextarea
-      v-model="elementData.question"
-      :readonly="isDisabled"
-      :rules="[requiredRule]"
-      class="my-3"
-      label="Question"
-      rows="3"
-      auto-grow
-    />
-    <div class="text-subtitle-2 mb-2">Select correct answers</div>
     <VInput
-      :model-value="!!elementData.correct.length"
-      :rules="[(val: any) => val || 'Please choose the correct answers']"
+      v-slot="{ isValid }"
+      :model-value="elementData.correct"
+      :rules="validation.correct"
     >
-      <div class="d-flex flex-column w-100">
-        <VSlideYTransition group>
-          <VTextField
-            v-for="(answer, index) in elementData.answers"
-            :key="index"
-            :model-value="answer"
-            :readonly="isDisabled"
-            :rules="[requiredRule]"
-            class="my-2"
-            placeholder="Answer..."
-            @update:model-value="updateAnswer(index, $event)"
-          >
-            <template #prepend>
-              <VCheckbox
-                v-model="elementData.correct"
-                :readonly="isDisabled"
-                :rules="[requiredRule]"
-                :validation-value="!!elementData.correct.length"
-                :value="index"
-                color="primary"
-                hide-details
-                multiple
-              />
-            </template>
-            <template #append>
-              <VBtn
-                v-if="!isDisabled && answersCount > 2"
-                aria-label="Remove answer"
-                density="comfortable"
-                icon="mdi-close"
-                variant="text"
-                @click="removeAnswer(index)"
-              />
-            </template>
-          </VTextField>
-        </VSlideYTransition>
-      </div>
+      <div class="text-subtitle-2 mb-2">{{ title }}</div>
+      <VSlideYTransition group>
+        <VTextField
+          v-for="(answer, index) in elementData.answers"
+          :key="index"
+          :model-value="answer"
+          :placeholder="placeholder"
+          :readonly="isDisabled"
+          :rules="validation.answer"
+          class="my-2 w-100"
+          variant="outlined"
+          @update:model-value="updateAnswer(index, $event)"
+        >
+          <template #prepend>
+            <VCheckbox
+              v-if="isGradable"
+              v-model="elementData.correct"
+              :error="isValid.value === false"
+              :readonly="isDisabled"
+              :value="index"
+              color="primary"
+              hide-details
+              multiple
+            />
+            <VAvatar
+              v-else
+              class="font-weight-bold ma-1"
+              color="primary-darken-3"
+              rounded="lg"
+              size="small"
+            >
+              {{ index + 1 }}
+            </VAvatar>
+          </template>
+          <template v-if="!isDisabled && answersCount > 2" #append>
+            <VBtn
+              aria-label="Remove answer"
+              color="primary-darken-4"
+              size="x-small"
+              variant="text"
+              icon
+              @click="removeAnswer(index)"
+            >
+              <VIcon icon="mdi-close" size="large" />
+            </VBtn>
+          </template>
+        </VTextField>
+      </VSlideYTransition>
     </VInput>
-    <div class="d-flex justify-center align-center mb-2">
+    <div class="d-flex justify-end mb-4">
       <VBtn
         v-if="!isDisabled"
+        color="primary-darken-4"
         prepend-icon="mdi-plus"
         variant="text"
         rounded
         @click="addAnswer"
       >
-        Add Answer
+        {{ btnLabel }}
       </VBtn>
     </div>
-    <div v-if="!isDisabled" class="d-flex justify-end">
-      <VBtn :disabled="isDirty" variant="text" @click="cancel">Cancel</VBtn>
-      <VBtn :disabled="isDirty" class="ml-2" type="submit" variant="tonal">
-        Save
-      </VBtn>
-    </div>
-  </VForm>
+  </QuestionContainer>
 </template>
 
 <script lang="ts" setup>
-import { computed, defineEmits, defineProps, reactive, ref, watch } from 'vue';
-import { Element, ElementData } from '@tailor-cms/ce-multiple-choice-manifest';
+import { computed, defineEmits, defineProps, reactive, watch } from 'vue';
+import manifest, {
+  Element,
+  ElementData,
+} from '@tailor-cms/ce-multiple-choice-manifest';
 import cloneDeep from 'lodash/cloneDeep';
 import isEqual from 'lodash/isEqual';
+import { QuestionContainer } from '@tailor-cms/core-components';
 
-const emit = defineEmits(['save']);
 const props = defineProps<{
+  embedTypes: any[];
   element: Element;
   isFocused: boolean;
   isDisabled: boolean;
 }>();
+const emit = defineEmits(['save']);
 
-const form = ref<HTMLFormElement>();
 const elementData = reactive<ElementData>(cloneDeep(props.element.data));
 
+const isGradable = computed(() => props.element.data.isGradable);
 const answersCount = computed(() => elementData.answers.length);
-const isDirty = computed(() => isEqual(elementData, props.element.data));
+const isDirty = computed(() => !isEqual(elementData, props.element.data));
+
+const title = computed(() =>
+  isGradable.value ? 'Select correct answer(s)' : 'Options',
+);
+const placeholder = computed(() =>
+  isGradable.value ? 'Answer...' : 'Option...',
+);
+const btnLabel = computed(() =>
+  isGradable.value ? 'Add answer' : 'Add option',
+);
+
+const validation = computed(() => ({
+  answer: [(val: string) => !!val || 'Answer is required'],
+  correct: isGradable.value
+    ? [(val?: number[]) => val?.length || 'Please choose the correct answer(s)']
+    : [],
+}));
 
 const addAnswer = () => elementData.answers.push('');
 const removeAnswer = (answerIndex: number) => {
   elementData.answers.splice(answerIndex, 1);
-  const index = elementData.correct.indexOf(answerIndex);
-  if (index !== -1) elementData.correct.splice(index, 1);
-  elementData.correct.forEach((it, i) => {
-    if (it >= answerIndex) elementData.correct[i] = it - 1;
-  });
+
+  if (isGradable.value) {
+    const index = elementData.correct!.indexOf(answerIndex);
+    if (index !== -1) elementData.correct!.splice(index, 1);
+    elementData.correct!.forEach((it, i) => {
+      if (it >= answerIndex && elementData.correct) {
+        elementData.correct[i] = it - 1;
+      }
+    });
+  }
 };
+
 const updateAnswer = (index: number, value: string) =>
   (elementData.answers[index] = value);
 
-const save = async () => {
-  const { valid } = await form.value?.validate();
-  if (valid) emit('save', elementData);
+const save = () => emit('save', elementData);
+
+const updateData = (data: ElementData) => {
+  Object.assign(elementData, cloneDeep(data));
 };
 
-const cancel = () => {
-  Object.assign(elementData, cloneDeep(props.element.data));
-  form.value?.resetValidation();
-};
-
-const requiredRule = (val: string | boolean | number) => {
-  return !!val || 'The field is required';
-};
-
-watch(
-  () => props.element.data,
-  (data) => Object.assign(elementData, cloneDeep(data)),
-);
+watch(() => props.element.data, updateData);
 </script>
-
-<style lang="scss" scoped>
-.tce-container {
-  text-align: left;
-}
-</style>