+++ packages/server/src/index.ts
@@ -1,9 +1,5 @@
 import type { HookServices, ServerRuntime } from '@tailor-cms/cek-common';
-import {
-  initState,
-  mocks,
-  type,
-} from '@tailor-cms/ce-multiple-choice-manifest';
+import { initState, type } from '@tailor-cms/ce-multiple-choice-manifest';
 import type { Element } from '@tailor-cms/ce-multiple-choice-manifest';
 import omit from 'lodash/omit.js';
 import xor from 'lodash/xor.js';
@@ -15,12 +11,10 @@ const USER_STATE: any = {};
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
 export function beforeSave(element: Element, services: HookServices) {
-  console.log('Before save hook');
   return element;
 }
 
 export function afterSave(element: Element, services: HookServices) {
-  console.log('After save hook');
   return element;
 }
 
@@ -29,7 +23,6 @@ export function afterLoaded(
   services: HookServices,
   runtime: ServerRuntime,
 ) {
-  console.log('After loaded hook');
   if (runtime === 'delivery') {
     const data = omit(element.data, ['correct']);
     return Object.assign(element, { data });
@@ -42,15 +35,13 @@ export function afterRetrieve(
   services: HookServices,
   runtime: ServerRuntime,
 ) {
-  console.log('After retrieve hook');
   return element;
 }
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export function beforeDisplay(element: Element, context: any) {
-  console.log('beforeDisplay hook');
-  console.log('beforeDisplay context', context);
-  return { ...context, ...USER_STATE, correct: element.data.correct };
+  if (IS_CEK) USER_STATE.correct = element.data.correct;
+  return { ...context, ...USER_STATE };
 }
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -59,12 +50,14 @@ export function onUserInteraction(
   context: any,
   payload: any,
 ): any {
-  console.log('onUserInteraction', context, payload);
+  const isGradable = element.data.isGradable;
   const isCorrect = xor(element.data.correct, payload.response).length === 0;
   // Simulate user state update within CEK
   if (IS_CEK) {
     // Can be reset to initial / mocked state via UI
-    Object.assign(context, { response: payload.response, isCorrect });
+    context.response = payload.response;
+    if (isGradable) context.isCorrect = isCorrect;
+    context.isSubmitted = true;
   }
   // Can have arbitrary return value (interpreted by target system)
   // FE is updated if updateDisplayState is true
@@ -92,7 +85,6 @@ export default {
   afterRetrieve,
   onUserInteraction,
   beforeDisplay,
-  mocks,
 };
 
-export { type, initState, mocks };
+export { type, initState };